requirements.txt
dj-database-url==0.5.0
Django==3.1.4
django-extensions==3.1.0
djangorestframework==3.12.2
docker==4.4.0
docker-compose==1.27.4
numpy==1.19.4
pandas==1.1.5
psycopg2-binary==2.8.3
gunicorn==20.0.4
whitenoise==5.2.0